refactor: move ArrayVisualizer and generateRandomArray to arrays_common.ts
